Изменен тип R + пофикшена первая четверть(полуокружность)

+++ src/main/java/jaba/web/fourthWebLab/ResultProcessing/CoordinateProcessing/CartesianAreas.java
@@ -1,9 +1,9 @@
 package jaba.web.fourthWebLab.ResultProcessing.CoordinateProcessing;
 
 public interface CartesianAreas {
-    boolean firstRotation(Double x, Double y, Float R);
-    boolean secondRotation(Double x, Double y, Float R);
-    boolean thirdRotation(Double x, Double y, Float R);
-    boolean fourthRotation(Double x,Double y,Float R);
-    boolean areaCheck(Double x,Double y,Float R);
+    boolean firstRotation(Double x, Double y, Double R);
+    boolean secondRotation(Double x, Double y, Double R);
+    boolean thirdRotation(Double x, Double y, Double R);
+    boolean fourthRotation(Double x,Double y,Double R);
+    boolean areaCheck(Double x, Double y, Double R);
 }
